qsorep add new commands: check, duplicate and query
The 'check' command can be used to check errors in ADIF log file fields.
However the 'adifmerg' does not check correct length of records (fields).

The 'duplicate' command tries to find duplicate QSOs from the log file.

The 'query' command can be used to make simple queries to the log file.
For example the list of DXCCs found from the file can be printed with
'file query dxcc'.
 # Please enter the commit message for your changes. Lines starting