adifmerg add new 3.0.7 fields and submodes
The data from www.adif.org version 3.0.7 was added to tables. The user
defined 'SOTA' and 'MY_SOTA' are replaced by 'SOTA_REF' and 'MY_SOTA_REF'
in ADIF 3.0.7.

Update list of bands and band limit frequencies.

Update list of propagation modes.

Create new hash 'submodes' which maps the submode to the mode family name.

Create new table 'modesimport' for modes that are not exported from
'adifmerg' (to be implemented). These modes are also in 'modes' table
and can be imported to 'adifmerg' for processing.