adifmerg routines to check and convert submodes
The 'CheckSubmode' checks if the submode is valid in ADIF 3.0.7.

When exporting ADIF file the MODE field is checked to see if it is a
submode in ADIF 3.0.7. If this is the case the content from MODE is
copied to SUBMODE field and MODE is set to the family name of these
modes. If SUBMODE is not empty no modification is done.