adifmerg handle QSL verified and manager cases
The 'QslVerified2Credit' changes the "V" flag to "Y" and adds new item
to CREDIT_GRANTED list. For example if QSL_RCVD was "V" new string
"AWARD:card" is added to CREDIT_GRANTED list. The name of the award
needs to be edited manually.

The "M" flag is depricated in QSL_RCVD_VIA and QSL_SENT_VIA. This flag
is replaced by empty string when exporting ADIF data.