adifmerg print SUBMODE if exists instead of MODE
To keep all the information in plain text and csv log files the
SUBMODE is printed instead of MODE. The SUBMODE can be used to
recover the MODE data if needed.

The '__MODE' tag in template (option '-T') files is replaced with SUBMODE
if exists. Thus the produced paper QSL cards will have most precise
definition of mode on them.