qsorep check QSO DXCC with option -c in adifmerg
The command 'file check dxcc [errors]' should be used only with local copy
of /usr/share/dxcc/cty.plist and not by calling 'dxcc'.

The printed errors need to be checked against received QSO confirmation which
could be a paper QSL card or record in the Logbook of the World. The file
'cty.plist' does not know where the contacted station was located when the
two way radio contact was made.