adifmerg add printing of county and state statistics with option -i
Total number of counties and states for all DXCC entities is printed.

With option -v list of counties and states is printed for each DXCC entity.

The county is also known as Secondary Administrative Subdivision in
adif.org and state as Primary Administrative Subdivision.

EQSL_QSL_RCVD and LOTW_QSL_RCVD flags 'Y' are now counted as confirmed
QSOs when printing number of DXCCs.